#3 adding startup filter to protect diagnostics endpoints
