+++ src/Prospa.Extensions.AspNetCore.Mvc.Core/StartupFilters/RequireEndpointKeyStartupFilter.cs
@@ -1,0 +1,58 @@
+﻿using System;
+using System.Linq;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Primitives;
+
+namespace Prospa.Extensions.AspNetCore.Mvc.Core.StartupFilters
+{
+    /// <summary>
+    /// Startup Filter to protect diagnostics endpoints such as health
+    /// </summary>
+    public class RequireEndpointKeyStartupFilter : IStartupFilter
+    {
+        private readonly string[] _endpoints;
+        private readonly string _key;
+
+        public RequireEndpointKeyStartupFilter(string[] endpoints, string key)
+        {
+            _endpoints = endpoints;
+            _key = key;
+        }
+
+        /// <inheritdoc />
+        public Action<IApplicationBuilder> Configure(Action<IApplicationBuilder> next)
+        {
+            return RequireSecretToMetricsAndHealth;
+
+            void RequireSecretToMetricsAndHealth(IApplicationBuilder app)
+            {
+                app.Use(async (context, next2) =>
+                {
+                    var key = ExtractToken(context);
+
+                    if (_endpoints.Any(e => string.Compare(context.Request.Path.Value, e, StringComparison.InvariantCultureIgnoreCase) == 0))
+                    {
+                        if (key != _key)
+                        {
+                            context.Abort();
+                            return;
+                        }
+                    }
+
+                    await next2.Invoke();
+                });
+
+                next(app);
+            }
+        }
+
+        private static string ExtractToken(HttpContext context)
+        {
+            return context.Request.QueryString.HasValue && context.Request.Query.ContainsKey("EndpointKey")
+                ? context.Request.Query["EndpointKey"]
+                : StringValues.Empty;
+        }
+    }
+}
